refactor: claim.ts and refund.ts
move from helpers
add refund.ts
add refundbutton test
swapchecker was interfering with status page
Update tests/components/RefundButton.spec.tsx
add testcase
