+++ src/pages/Refund.tsx
@@ -4,6 +4,7 @@ import QrScanner from "qr-scanner";
 import { Show, createEffect, createSignal } from "solid-js";
 
 import BlockExplorer from "../components/BlockExplorer";
+import RefundButton from "../components/RefundButton";
 import RefundEta from "../components/RefundEta";
 import SwapList from "../components/SwapList";
 import t from "../i18n";
@@ -11,35 +12,23 @@ import {
     refundTx,
     setTimeoutBlockheight,
     setTimeoutEta,
-    setTransactionToRefund,
     swaps,
 } from "../signals";
-import { fetcher, refund, refundAddressChange } from "../utils/helper";
+import { fetcher } from "../utils/helper";
+import { refundJsonKeys, refundJsonKeysLiquid } from "../utils/refund";
 import {
     swapStatusFailed,
     swapStatusSuccess,
     updateSwapStatus,
 } from "../utils/swapStatus";
 
-const refundJsonKeys = ["id", "asset", "privateKey", "redeemScript"];
-const refundJsonKeysLiquid = refundJsonKeys.concat("blindingKey");
-
 const Refund = () => {
     const navigate = useNavigate();
 
-    const [valid, setValid] = createSignal(false);
-    const [addressValid, setAddressValid] = createSignal(false);
-    const [refundJsonValid, setRefundJsonValid] = createSignal(false);
-    const [refundable, setRefundable] = createSignal(true);
     const [refundJson, setRefundJson] = createSignal(null);
 
-    createEffect(() => {
-        if (addressValid() && refundJsonValid()) {
-            setValid(true);
-        } else {
-            setValid(false);
-        }
-    });
+    setTimeoutBlockheight(null);
+    setTimeoutEta(null);
 
     const checkRefundJsonKeys = (input: HTMLInputElement, json: any) => {
         log.debug("checking refund json", json);
@@ -59,8 +48,6 @@ const Refund = () => {
             json.asset !== "L-BTC" ? refundJsonKeys : refundJsonKeysLiquid;
         const valid = requiredKeys.every((key) => key in json);
 
-        setRefundJsonValid(valid);
-
         if (valid) {
             setRefundJson(json);
         } else {
@@ -72,8 +59,7 @@ const Refund = () => {
         const input = e.currentTarget as HTMLInputElement;
         const inputFile = input.files[0];
         input.setCustomValidity("");
-        setRefundJson("");
-        setRefundJsonValid(false);
+        setRefundJson(null);
 
         if (inputFile.type === "image/png") {
             QrScanner.scanImage(inputFile, { returnDetailedScanResult: true })
@@ -93,30 +79,6 @@ const Refund = () => {
                     input.setCustomValidity(t("invalid_refund_file"));
                 });
         }
-    };
-
-    const startRefund = () => {
-        if (!valid()) return;
-        const refundInfo = refundJson();
-        fetcher(
-            "/getswaptransaction",
-            refundInfo.asset,
-            async (data: any) => {
-                if (data.timeoutEta) {
-                    setTimeoutEta(data.timeoutEta * 1000);
-                    setTimeoutBlockheight(data.timeoutBlockHeight);
-                    setRefundable(false);
-                    return;
-                }
-
-                setRefundable(true);
-                setTransactionToRefund(data);
-                await refund(refundJson(), t);
-            },
-            {
-                id: refundInfo.id,
-            },
-        );
     };
 
     const refundSwapsSanityFilter = (swap: any) =>
@@ -200,27 +162,8 @@ const Refund = () => {
                     accept="application/json,image/png"
                     onChange={(e) => uploadChange(e)}
                 />
-                <input
-                    required
-                    disabled={!refundJsonValid()}
-                    onInput={(e) =>
-                        setAddressValid(
-                            refundAddressChange(e, refundJson().asset),
-                        )
-                    }
-                    type="text"
-                    id="refundAddress"
-                    name="refundAddress"
-                    placeholder={t("refund_address_placeholder")}
-                />
-                <button
-                    class="btn"
-                    disabled={valid() && refundTx() === ""}
-                    onClick={startRefund}>
-                    {t("refund")}
-                </button>
-                <Show when={!refundable()}>
-                    <hr />
+                <Show when={refundTx() === ""}>
+                    <RefundButton swap={refundJson} />
                     <RefundEta />
                 </Show>
                 <Show when={refundTx() !== ""}>
